Proposed changes
Overall, it's really great and so catchy!

I rearranged a few items where I thought it was unclear. I added some images I found from past years. Maybe someone else can find a couple more?

I also propose moving the deadline to April 1. I think this would be a great way to keep people engaged after the end of the call for papers. If we announce on Feb 1. That would give participants 2 months to submit something and gives us enough time to make a decision and modify the images before we need to use the labels.

I also changed the hashtag to GR8beer since I think the other one is too long and might be easily confused with BlackLabel. I don't think we'll get that many submissions that the hashtag is flooded.

Open items:
- It would be nice to have a defined prize before we announce the contest. I'm not sure how many people will submit if they don't know if they'll be at the conference. I think a conference ticket (without travel) or at least a discounted ticket would be fair.